requirements.txt
glob
cv2
numpy
face_recognition
pillow
pillow_heif
streamlit